build compile fixes

+++ app/components/CommentsSection.tsx
@@ -1,5 +1,4 @@
-import React, { FC } from "react";
-import { useState, useEffect } from "react";
+import React, { FC, useRef, useState, useEffect } from "react";
 import { useSession } from "next-auth/react";
 import { comments as commentsType } from "../.../../../types";
 import {
@@ -30,34 +29,44 @@ type CommentsSectionProps = {
 const CommentsSection: FC<CommentsSectionProps> = ({ comments, id }) => {
   const [commentsList, setCommentsList] = useState<commentsType[]>([]);
   const [deleteWarning, setDeleteWarning] = useState(false);
-  const [deletedID, setDeletedID] = useState("");
+  const [deletedID, setDeletedID] = useState<string | null>(null);
   const [deleteSuccess, setDeleteSuccess] = useState(false);
   const [deleteError, setDeleteError] = useState(false);
 
   const { isOpen, onOpen, onClose } = useDisclosure();
-  const cancelRef = React.useRef();
+  const cancelRef = useRef<HTMLButtonElement>(null);
   const { data: session } = useSession();
 
   const deleteComment = async () => {
     setDeleteWarning(false);
-    const response = await fetch(`/api/forums/${id}/comments/${deletedID}`, {
-      method: "DELETE",
-    });
-    if (response.ok) {
-      setCommentsList(commentsList.filter((comment: commentsType) => comment._id.toString() !== deletedID));
-
-      setDeleteSuccess(true);
-    } else {
-      setDeleteError(true);
+    if (deletedID) {
+      const response = await fetch(`/api/forums/${id}/comments/${deletedID}`, {
+        method: "DELETE",
+      });
+      if (response.ok) {
+        setCommentsList(
+          commentsList.filter(
+            (comment: commentsType) => comment._id.toString() !== deletedID
+          )
+        );
+        setDeleteSuccess(true);
+      } else {
+        setDeleteError(true);
+      }
     }
   };
 
   useEffect(() => {
     setCommentsList(comments);
-  }, []);
+  }, [comments]);
+
   return (
     <Flex flexDirection="column" gap="3">
-      <PostComment itemId={id} setCommentsList={setCommentsList} commentList={commentsList} />
+      <PostComment
+        itemId={id}
+        setCommentsList={setCommentsList}
+        commentList={commentsList}
+      />
       {deleteSuccess && (
         <Flex flexDirection="column" alignItems="center">
           <Text>Comment deleted</Text>{" "}
@@ -86,22 +95,31 @@ const CommentsSection: FC<CommentsSectionProps> = ({ comments, id }) => {
           </Button>
         </Flex>
       )}
-      <Heading fontSize="xl">{commentsList.length} Comments </Heading>
+      <Heading fontSize="xl">{commentsList.length} Comments</Heading>
       <Grid templateColumns={{ base: "1fr", md: "repeat(4, 1fr)" }} gap={6}>
-        {commentsList.map((comment: any) => {
+        {commentsList.map((comment: commentsType) => {
           const currentDate = new Date();
           const commentDate = new Date(comment.date);
-          let timeSincePost
-          if(currentDate.getTime() - commentDate.getTime() < 86400000){
-           timeSincePost = (Math.round(((currentDate.getTime() - commentDate.getTime())) / 36000000).toString() + 'h')
-          }
-          else{
-          timeSincePost = (Math.round(((currentDate.getTime() - commentDate.getTime())) / 8.64e7).toString() + 'd')
-          }
+          const timeSincePost =
+            currentDate.getTime() - commentDate.getTime() < 86400000
+              ? Math.round(
+                  (currentDate.getTime() - commentDate.getTime()) / 3600000
+                ) + "h"
+              : Math.round(
+                  (currentDate.getTime() - commentDate.getTime()) / 86400000
+                ) + "d";
+
           return (
-            <GridItem bg="none">
+            <GridItem bg="none" key={comment._id.toString()}>
+              {" "}
               <Text fontSize="sm">{comment.body}</Text>
-              <Flex flexDirection="row" justifyContent="space-between" alignItems="center" mt="2" fontSize="sm">
+              <Flex
+                flexDirection="row"
+                justifyContent="space-between"
+                alignItems="center"
+                mt="2"
+                fontSize="sm"
+              >
                 <Stack direction="row">
                   <Text fontStyle="italic" fontWeight="bold">
                     {comment.author}
@@ -115,38 +133,56 @@ const CommentsSection: FC<CommentsSectionProps> = ({ comments, id }) => {
                       colorScheme="red"
                       variant="outline"
                       onClick={() => {
-                        setDeleteWarning(true), setDeletedID(comment._id);
+                        setDeletedID(comment._id.toString());
+                        setDeleteWarning(true);
+                        onOpen();
                       }}
                     >
                       Delete
                     </Button>
                   </Box>
                 )}
-
-                <AlertDialog isOpen={deleteWarning} leastDestructiveRef={cancelRef} onClose={onClose}>
+                <AlertDialog
+                  isOpen={deleteWarning}
+                  leastDestructiveRef={cancelRef}
+                  onClose={() => {
+                    setDeleteWarning(false);
+                    onClose();
+                  }}
+                >
                   <AlertDialogOverlay>
                     <AlertDialogContent>
                       <AlertDialogHeader fontSize="lg" fontWeight="bold">
                         Delete Comment
                       </AlertDialogHeader>
-                      <AlertDialogBody>Are you sure? You can't undo this action afterwards.</AlertDialogBody>
+                      <AlertDialogBody>
+                        Are you sure? You can&#39;t undo this action afterwards.
+                      </AlertDialogBody>
                       <AlertDialogFooter>
                         <Button
                           ref={cancelRef}
                           onClick={() => {
                             setDeleteWarning(false);
+                            onClose();
                           }}
                         >
                           Cancel
                         </Button>
-                        <Button colorScheme="red" onClick={deleteComment} ml={3}>
+                        <Button
+                          colorScheme="red"
+                          onClick={deleteComment}
+                          ml={3}
+                        >
                           Delete
                         </Button>
                       </AlertDialogFooter>
                     </AlertDialogContent>
                   </AlertDialogOverlay>
                 </AlertDialog>
-                <LikeDislikeButtons itemId={comment._id} type={"forums/" + id + "/comments"} />
+                <LikeDislikeButtons
+                  itemId={comment._id.toString()}
+                  type={`forums/${id}/comments`}
+                />{" "}
               </Flex>
             </GridItem>
           );
